backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.7.0
torchvision==0.22.0
timm==0.9.7
pillow==10.1.0
numpy==1.26.2 